newsfeed/requirements.txt
httpx>=0.27.0
python-dotenv>=1.0.0
google-genai>=1.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.13.0
python-fasthtml>=0.12.0
MonsterUI
